Logout with cancel

+++ app/routes/logout.tsx
@@ -1,5 +1,6 @@
 import {type ActionArgs, type LoaderArgs, redirect} from "@remix-run/node";
 import {Form} from "@remix-run/react";
+import invariant from "tiny-invariant";
 
 import {PageWrapper} from "~/components/common/page_wrapper";
 import {H1} from "~/components/ui/typography";
@@ -16,22 +17,38 @@ export async function loader({request}: LoaderArgs) {
 }
 
 export async function action({request}: ActionArgs) {
-  let session = await getWJSSession(request.headers.get("Cookie"));
-  return redirect("/", {
-    headers: {
-      "Set-Cookie": await destroyWJSSession(session),
-    },
-  });
+  let form = await request.formData();
+  let logout = form.get("logout");
+  invariant(logout !== null, "logout should not be null");
+  invariant(typeof logout === "string", "logout should be a string");
+  if (logout === "yes") {
+    let session = await getWJSSession(request.headers.get("Cookie"));
+    return redirect("/", {
+      headers: {
+        "Set-Cookie": await destroyWJSSession(session),
+      },
+    });
+  }
+  return redirect("/");
 }
 
 export default function Logout() {
   return (
     <PageWrapper>
-      <H1>Logout</H1>
-      <p>Are you sure you want to logout?</p>
-      <Form method="post">
-        <Button type="submit">Logout</Button>
-      </Form>
+      <section className="flex flex-col gap-2  pt-10">
+        <H1>Logout</H1>
+        <p>Are you sure you want to logout?</p>
+        <Form method="post">
+          <div className="flex gap-2">
+            <Button variant="primary" name="logout" value="yes" type="submit">
+              Logout
+            </Button>
+            <Button name="logout" value="no" type="submit">
+              Cancel
+            </Button>
+          </div>
+        </Form>
+      </section>
     </PageWrapper>
   );
 }
